begin refactoring of study page
