+++ packages/frontend/src/features/Study/StudyDetails/RendererFactory.tsx
