+++ packages/frontend/src/features/Study/StudyDetails/StudyDetailsPanel.tsx
