+++ packages/frontend/src/features/Study/StudyDetails/StudyGroupPanel.tsx
