+++ packages/frontend/src/features/Study/StudyPage/StudyPage.tsx
