+++ packages/frontend/src/features/Study/index.ts
@@ -1,0 +1,5 @@
+import StudyGroup from './StudyDetails/StudyGroup';
+
+export {
+  StudyGroup,
+}
