+++ packages/frontend/src/utils/access.ts
@@ -1,0 +1,21 @@
+import {
+  JSONValue,
+  type MetadataPaginationParams,
+  type AggregationsData,
+} from '@gen3/core';
+
+export const accessibleFieldName = '__accessible';
+
+export enum AccessLevel {
+  ACCESSIBLE = 1,
+  UNACCESSIBLE = 2,
+  WAITING = 3,
+  NOT_AVAILABLE = 4,
+  OTHER = 5,
+  MIXED = 6,
+}
+
+export interface AccessibleResource
+  extends Record<string, JSONValue | AccessLevel | undefined> {
+  [accessibleFieldName]?: AccessLevel;
+}
