+++ packages/frontend/src/utils/dataAuthorization.ts
@@ -1,0 +1,11 @@
+export interface AuthorizationValues {
+  enabled: boolean;
+  menuText: string;
+}
+
+export interface DataAuthorization {
+  columnTooltip?: string;
+  supportedValues?: Record<string, AuthorizationValues>;
+  isMesh?: boolean;
+  enabled?: boolean;
+}
