+++ packages/frontend/src/utils/index.ts
@@ -11,3 +11,5 @@ export * from './strings';
 export * from './isType';
 export * from './values';
 export * from './time';
+export * from './access';
+export * from './dataAuthorization';
